% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boids.R
\name{new_Boids}
\alias{new_Boids}
\title{Constructor Boid}
\usage{
new_Boids(
  pos_x = double(),
  pos_y = double(),
  vel_x = double(),
  vel_y = double()
)
}
\arguments{
\item{vel_y}{}
}
\value{

}
\description{
Constructor Boid
}
